+++ app/(home)/contents.module.scss
@@ -63,6 +63,61 @@
   padding: 0 0.05em;
   line-height: 100%;
 }
+.sns {
+  margin-top: 2rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.snsIcon {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  > a {
+    margin-left: 2rem;
+    margin-right: 2rem;
+  }
+}
+
+%icon {
+  font-size: 2.5rem;
+  color: $color-darker;
+  transition: color 0.3s ease;
+}
+
+.youtube {
+  @extend %icon;
+
+  &:hover {
+    color: $color-youtube;
+  }
+}
+
+.twitter {
+  @extend %icon;
+
+  &:hover {
+    color: $color-twitter;
+  }
+}
+
+.instagram {
+  @extend %icon;
+
+  &:hover {
+    color: $color-instagram;
+  }
+}
+
+.line {
+  @extend %icon;
+
+  &:hover {
+    color: $color-line;
+  }
+}
 
 .slogan {
   @include container();
