+++ app/(home)/page.module.scss
@@ -1,5 +1,4 @@
 @use '@variables' as *;
-$color-darker: #333;
 
 .main {
   display: flex;
@@ -28,60 +27,4 @@ $color-darker: #333;
   letter-spacing: 0.1rem;
   margin-top: 0;
   color: $color-darker;
-}
-
-.sns {
-  margin-top: 2rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.snsIcon {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  > a {
-    margin-left: 2rem;
-    margin-right: 2rem;
-  }
-}
-
-%icon {
-  font-size: 2.5rem;
-  color: $color-darker;
-  transition: color 0.3s ease;
-}
-
-.youtube {
-  @extend %icon;
-
-  &:hover {
-    color: $color-youtube;
-  }
-}
-
-.twitter {
-  @extend %icon;
-
-  &:hover {
-    color: $color-twitter;
-  }
-}
-
-.instagram {
-  @extend %icon;
-
-  &:hover {
-    color: $color-instagram;
-  }
-}
-
-.line {
-  @extend %icon;
-
-  &:hover {
-    color: $color-line;
-  }
 }
