+++ app/variables.scss
@@ -1,3 +1,4 @@
+$color-darker: #333;
 $color-white: #ffffff;
 $color-red: #e53242;
 $color-purple: #9272af;
